+++ CHANGELOG.md
@@ -4,6 +4,17 @@ We are following [Semantic versioning](https://semver.org/) in this library
 
 # Changes in releases
 
+## pre-release 0.2.0
+
+Fixes / Breaking (early):
+
+- All `kt_precond` methods return `*kt_errors.FaultBuilder` instead of a `kt_errors.Fault` so we can tweak further the Fault before we build and throw it.
+
+New features:
+
+- Added `EnsureState()` and `EnsureStatePub` methods to check states and throw `kt_errors.IllegalStateFault` if condition not met.
+- Added quick example code to README.md
+
 ## pre-release 0.1.0
 
 Initial commit - quick release to start tasting it and validate if it is helpful.